feat: move all files to them specific packages for better organization and grant clean artchitecture
