+++ bills/models/delete_bills.go
