+++ bills/models/get_all_bills.go
