+++ bills/models/get_bills.go
