+++ budgets/handlers/handler_delete.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"strconv"
 
-	models "github.com/Dnreikronos/budgetMannager---Back/models/budgets"
+	"github.com/Dnreikronos/budgetMannager---Back/budgets/models"
 )
 
 func DeleteBudgets(w http.ResponseWriter, r *http.Request) {
