+++ users/handlers/handler_delete.go
@@ -6,11 +6,10 @@ import (
 	"net/http"
 	"strconv"
 
-	models "github.com/Dnreikronos/budgetMannager---Back/models/users"
+	"github.com/Dnreikronos/budgetMannager---Back/users/models"
 )
 
 func deleteUser(w http.ResponseWriter, r *http.Request) {
-
 	if r.Method != http.MethodDelete {
 		http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
 		return
