+++ users/handlers/handler_get_all.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"net/http"
 
-	models "github.com/Dnreikronos/budgetMannager---Back/models/users"
+	"github.com/Dnreikronos/budgetMannager---Back/users/models"
 )
 
 func getAllUser(w http.ResponseWriter, r *http.Request) {
